requirements.txt
Package             Version
------------------- ----------
asgiref             3.2.7
autopep8            1.5.3
backcall            0.2.0
certifi             2020.4.5.2
chardet             3.0.4
decorator           4.4.2
dj-database-url     0.5.0
Django              3.0.7
django-cors-headers 3.3.0
django-filter       2.3.0
django-heroku       0.3.1
djangorestframework 3.11.0
gunicorn            20.0.4
idna                2.9
importlib-metadata  1.7.0
ipython             7.16.1
ipython-genutils    0.2.0
jedi                0.17.1
Markdown            3.2.2
parso               0.7.0
pexpect             4.8.0
pickleshare         0.7.5
pip                 20.1
prompt-toolkit      3.0.5
psycopg2            2.8.5
ptyprocess          0.6.0
pycodestyle         2.6.0
Pygments            2.6.1
pytz                2020.1
requests            2.23.0
setuptools          46.1.3
six                 1.15.0
sqlparse            0.3.1
toml                0.10.1
traitlets           4.3.3
urllib3             1.25.9
wcwidth             0.2.5
wheel               0.34.2
whitenoise          5.1.0
zipp                3.1.0
